refactor: 페이지 상태 로직 분리 및 섹션 컴포넌트화
